# classroom/apps/go/internal/classroom/docs/swagger.yaml
basePath: /api
definitions:
  auth.TokenDetails:
    properties:
      accessExpires:
        description: AccessExpires Access token expiration
        type: integer
      accessToken:
        type: string
      accessUUID:
        description: Unique identifier for access token
        type: string
      refreshExpires:
        description: RefreshExpires Refresh token expiration
        type: integer
      refreshToken:
        type: string
      refreshUUID:
        description: Unique identifier for refresh token
        type: string
    type: object
...
  handler.AuthResponse:
    properties:
      token:
        $ref: '#/definitions/auth.TokenDetails'
      user:
        $ref: '#/definitions/schema.User'
    type: object
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  model.Comment:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      post:
        $ref: '#/definitions/model.Post'
      postId:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
  model.File:
    properties:
      bucket:
        type: string
      createdAt:
        type: string
      fileableId:
        type: string
      fileableType:
        type: string
      id:
        type: integer
      key:
        type: string
      mimeType:
        type: string
      type:
        type: integer
      updatedAt:
        type: string
      url:
        type: string
    type: object
  model.Post:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      content:
        type: string
      createdAt:
        type: string
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      id:
        type: integer
      sectionId:
        type: integer
      type:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      id:
        type: string
      lastname:
        type: string
      name:
        type: string
      profileImage:
        $ref: '#/definitions/model.File'
      updatedAt:
        type: string
    type: object
...
  request.CommentPostRequest:
    properties:
      content:
        type: string
      post:
        type: integer
      userId:
        type: string
    required:
    - content
    - post
    type: object
...
  request.StorePostRequest:
    properties:
      content:
        type: string
      section:
        type: integer
      userID:
        type: string
    required:
    - content
    - section
    - userID
    type: object
...
  schema.User:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
host: localhost
info:
  contact:
    name: Binaretech
  description: Main API for the classroom. Full source code at https://github.com/Binaretech/classroom
  license:
    name: Apache 2.0
  title: Classroom API
  version: "0.1"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login the user
      parameters:
      - description: Login data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthResponse'
      summary: Login
      tags:
      - Auth
...
  /posts/:id/comments:
    post:
      consumes:
      - application/json
      description: Comment a post
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/request.CommentPostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Comment'
      summary: Comment a post
      tags:
      - Posts
  /register:
    get:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.AuthResponse'
      summary: Register a new user
      tags:
      - Auth
...
  /sections/:id/posts:
    post:
      description: Store a post in the database
      parameters:
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/request.StorePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Post'
      summary: Store a post
      tags:
      - Posts
...
  /sections/{id}/posts:
    get:
      consumes:
      - application/json
      description: Get section posts
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
      summary: Get section posts
      tags:
      - Sections
...
securityDefinitions:
  Authorization Token:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"